migrate file actions to depositForm actions
