+++ src/main/typescript/actions/depositFormActions.ts
@@ -17,7 +17,6 @@ import * as H from "history"
 import { DepositFormMetadata } from "../components/form/parts"
 import { ComplexThunkAction, FetchAction, PromiseAction, ReduxAction, ThunkAction } from "../lib/redux"
 import { DepositFormConstants } from "../constants/depositFormConstants"
-import { FileOverviewConstants } from "../constants/fileOverviewConstants"
 import { DepositId } from "../model/Deposits"
 import { Action } from "redux"
 import { AppState } from "../model/AppState"
@@ -26,6 +25,7 @@ import { Doi } from "../lib/metadata/Identifier"
 import { filesConverter } from "../lib/files/files"
 import fetch from "../lib/fetch"
 import {
+    deleteFileUrl,
     fetchDepositStateUrl,
     fetchDoiUrl,
     fetchMetadataUrl,
@@ -34,6 +34,7 @@ import {
     submitDepositUrl,
 } from "../selectors/serverRoutes"
 import { DepositState } from "../model/DepositState"
+import { Files } from "../model/FileInfo"
 import { depositStateConverter } from "../lib/depositState/depositState"
 import { AxiosResponse } from "axios"
 
@@ -56,6 +57,17 @@ export const depositStateNotFound: () => Action = () => ({
     type: DepositFormConstants.FETCH_STATE_NOT_FOUND,
 })
 
+export const fetchFiles: (depositId: DepositId) => ThunkAction<FetchAction<Files>> = (depositId) => (dispatch, getState) => dispatch({
+    type: DepositFormConstants.FETCH_FILES,
+    async payload() {
+        const response = await fetch.get(listFilesUrl(depositId)(getState()))
+        return response.data
+    },
+    meta: {
+        transform: filesConverter,
+    },
+})
+
 export const fetchMetadata: (depositId: DepositId) => ThunkAction<FetchAction<DepositFormMetadata, AppState>> = depositId => (dispatch, getState) => dispatch({
     type: DepositFormConstants.FETCH_METADATA,
     async payload() {
@@ -72,7 +84,7 @@ export const fetchMetadataAndFiles: (depositId: DepositId) => ComplexThunkAction
     const metadataPromise = fetch.get(fetchMetadataUrl(depositId)(getState()))
 
     dispatch({
-        type: FileOverviewConstants.FETCH_FILES_PENDING,
+        type: DepositFormConstants.FETCH_FILES_PENDING,
     })
     dispatch({
         type: DepositFormConstants.FETCH_METADATA_PENDING,
@@ -84,20 +96,20 @@ export const fetchMetadataAndFiles: (depositId: DepositId) => ComplexThunkAction
             const fileResponse = await filePromise
 
             dispatch({
-                type: FileOverviewConstants.FETCH_FILES_FULFILLED,
+                type: DepositFormConstants.FETCH_FILES_FULFILLED,
             })
             dispatch({
                 type: DepositFormConstants.FETCH_METADATA_SUCCESS,
                 payload: {
                     ...(metadataConverter(metadataResponse.data, getState().dropDowns)),
                     files: filesConverter(fileResponse.data),
-                }
+                },
             })
         }
         catch (fileError) {
             dispatch({
-                type: FileOverviewConstants.FETCH_FILES_REJECTED,
-                payload: fileError
+                type: DepositFormConstants.FETCH_FILES_REJECTED,
+                payload: fileError,
             })
         }
     }
@@ -201,4 +213,30 @@ export const submitDeposit: (depositId: DepositId, data: DepositFormMetadata, hi
 export const submitDepositRejectedAction: (errorMessage: string) => ReduxAction<string> = errorMessage => ({
     type: DepositFormConstants.SUBMIT_DEPOSIT_REJECTED,
     payload: errorMessage,
+})
+
+export const deleteFile: (depositId: DepositId, filePath: string, setToDraft: boolean) => ThunkAction<PromiseAction<void>> = (depositId, filePath, setToDraft) => (dispatch, getState) => dispatch({
+    type: DepositFormConstants.DELETE_FILE,
+    async payload() {
+        if (setToDraft)
+            await setStateToDraft(depositId, getState)
+
+        await fetch.delete(deleteFileUrl(depositId, filePath)(getState()))
+        dispatch(fetchFiles(depositId))
+    },
+    meta: {
+        filePath: filePath,
+        depositId: depositId,
+        setStateToDraft: setToDraft,
+    },
+})
+
+export const cancelDeleteFile: (filePath: string) => Action = (filePath) => ({
+    type: DepositFormConstants.DELETE_FILE_CANCELLED,
+    meta: { filePath: filePath },
+})
+
+export const askConfirmationToDeleteFile: (filePath: string) => Action = (filePath) => ({
+    type: DepositFormConstants.DELETE_FILE_CONFIRMATION,
+    meta: { filePath: filePath },
 })
