--- src/main/typescript/actions/fileOverviewActions.ts
@@ -1,61 +1,0 @@
-/*
- * Copyright (C) 2018 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { FetchAction, PromiseAction, ThunkAction } from "../lib/redux"
-import { FileOverviewConstants } from "../constants/fileOverviewConstants"
-import { Action } from "redux"
-import { DepositId } from "../model/Deposits"
-import { filesConverter } from "../lib/files/files"
-import fetch from "../lib/fetch"
-import { deleteFileUrl, listFilesUrl } from "../selectors/serverRoutes"
-import { Files } from "../model/FileInfo"
-import { setStateToDraft } from "./depositFormActions"
-
-export const fetchFiles: (depositId: DepositId) => ThunkAction<FetchAction<Files>> = (depositId) => (dispatch, getState) => dispatch({
-    type: FileOverviewConstants.FETCH_FILES,
-    async payload() {
-        const response = await fetch.get(listFilesUrl(depositId)(getState()))
-        return response.data
-    },
-    meta: {
-        transform: filesConverter,
-    },
-})
-
-export const deleteFile: (depositId: DepositId, filePath: string, setToDraft: boolean) => ThunkAction<PromiseAction<void>> = (depositId, filePath, setToDraft) => (dispatch, getState) => dispatch({
-    type: FileOverviewConstants.DELETE_FILE,
-    async payload() {
-        if (setToDraft)
-            await setStateToDraft(depositId, getState)
-
-        await fetch.delete(deleteFileUrl(depositId, filePath)(getState()))
-        dispatch(fetchFiles(depositId))
-    },
-    meta: {
-        filePath: filePath,
-        depositId: depositId,
-        setStateToDraft: setToDraft,
-    },
-})
-
-export const cancelDeleteFile: (filePath: string) => Action = (filePath) =>({
-    type: FileOverviewConstants.DELETE_FILE_CANCELLED,
-    meta: { filePath: filePath },
-})
-
-export const askConfirmationToDeleteFile: (filePath: string) => Action = (filePath) =>({
-    type: FileOverviewConstants.DELETE_FILE_CONFIRMATION,
-    meta: { filePath: filePath },
-})
