+++ src/main/typescript/components/form/parts/fileUpload/overview/FilesOverview.tsx
@@ -23,7 +23,7 @@ import { DepositState } from "../../../../../model/DepositState"
 import { emptyFiles, Files } from "../../../../../model/FileInfo"
 import { useSelector } from "../../../../../lib/redux"
 import { FieldProps } from "../../../../../lib/formComponents/ReduxFormUtils"
-import { askConfirmationToDeleteFile, cancelDeleteFile, deleteFile } from "../../../../../actions/fileOverviewActions"
+import { askConfirmationToDeleteFile, cancelDeleteFile, deleteFile } from "../../../../../actions/depositFormActions"
 import { CloseableWarning } from "../../../../Errors"
 import Loading from "../../../../Loading"
 import Paginationable from "../../../../Paginationable"
