+++ src/main/typescript/constants/depositFormConstants.ts
@@ -23,6 +23,12 @@ export enum DepositFormConstants {
     FETCH_STATE_SUCCESS = "FETCH_STATE_SUCCESS",
     FETCH_STATE_NOT_FOUND = "FETCH_STATE_NOT_FOUND",
 
+    FETCH_FILES = "FETCH_FILES",
+    FETCH_FILES_PENDING = "FETCH_FILES_PENDING",
+    FETCH_FILES_REJECTED = "FETCH_FILES_REJECTED",
+    FETCH_FILES_FULFILLED = "FETCH_FILES_FULFILLED",
+    FETCH_FILES_SUCCESS = "FETCH_FILES_SUCCESS",
+
     FETCH_METADATA = "FETCH_METADATA",
     FETCH_METADATA_PENDING = "FETCH_METADATA_PENDING",
     FETCH_METADATA_FULFILLED = "FETCH_METADATA_FULFILLED",
@@ -45,6 +51,13 @@ export enum DepositFormConstants {
     SUBMIT_DEPOSIT_PENDING = "SUBMIT_DEPOSIT_PENDING",
     SUBMIT_DEPOSIT_FULFILLED = "SUBMIT_DEPOSIT_FULFILLED",
     SUBMIT_DEPOSIT_REJECTED = "SUBMIT_DEPOSIT_REJECTED",
+
+    DELETE_FILE_CONFIRMATION = "DELETE_FILE_CONFIRMATION",
+    DELETE_FILE_CANCELLED = "DELETE_FILE_CANCELLED",
+    DELETE_FILE = "DELETE_FILE",
+    DELETE_FILE_PENDING = "DELETE_FILE_PENDING",
+    DELETE_FILE_FULFILLED = "DELETE_FILE_FULFILLED",
+    DELETE_FILE_REJECTED = "DELETE_FILE_REJECTED",
 }
 
 export const depositFormName = "depositForm"
