+++ src/main/typescript/middleware/depositFormMiddleware.ts
@@ -19,7 +19,6 @@ import { depositOverviewRoute } from "../constants/clientRoutes"
 import { depositStateNotFound, fetchDepositState, saveDraftResetAction } from "../actions/depositFormActions"
 import { actionTypes, change, initialize } from "redux-form"
 import { DAI, ISNI, ORCID } from "../lib/metadata/Contributor"
-import { FileOverviewConstants } from "../constants/fileOverviewConstants"
 
 const depositStateNotFoundMiddleware: Middleware = ({ dispatch }) => (next: Dispatch) => action => {
     if (action.type && action.type === DepositFormConstants.FETCH_STATE_REJECTED && action.payload) {
@@ -44,14 +43,14 @@ const fetchDoiProcessor: Middleware = ({ dispatch }: MiddlewareAPI) => (next: Di
 const fetchFilesProcessor: Middleware = ({ dispatch }: MiddlewareAPI) => (next: Dispatch) => action => {
     next(action)
 
-    if (action.type === FileOverviewConstants.FETCH_FILES_SUCCESS)
+    if (action.type === DepositFormConstants.FETCH_FILES_SUCCESS)
         dispatch(change(depositFormName, "files", action.payload))
 }
 
 const fetchStateAfterSetFromRejectedToDraft: Middleware = ({ dispatch }: MiddlewareAPI<Dispatch<any>>) => (next: Dispatch) => action => {
     next(action)
 
-    if ((action.type === DepositFormConstants.SAVE_DRAFT_FULFILLED || action.type === FileOverviewConstants.DELETE_FILE_FULFILLED)
+    if ((action.type === DepositFormConstants.SAVE_DRAFT_FULFILLED || action.type === DepositFormConstants.DELETE_FILE_FULFILLED)
         && action.meta?.setStateToDraft
         && action.meta?.depositId)
         dispatch(fetchDepositState(action.meta.depositId))
